Ensure PROC/FN pointers are NULL after load.
Prior to using the end of page for the program code itself the pointers
for the PROC and FN lists were not touched but since loading program
code in that area this would leave them as dangling pointers.  This
ensures zeros are loaded to those locations so the lists are correct
empty lists.